Notebooks/Part1/requirements.txt
#Scraping
selenium
webdriver_manager
beautifulsoup4
pandas
boto3
botocore
six
python-dotenv

#pinecone
unstructured
tiktoken
pinecone-client
pypdf
openai==0.28.0
langchain
numpy
python-dotenv
streamlit

